pyproject.toml
[project]
name = "CliRenderer"
version = "0.1.8"
requires-python = ">=3.9"
authors = [
    { name = "ultraflame4", email = "ultraflame4@gmail.com" }
]
description = "A CLI tool to render unicode art in the terminal"
readme = "README.md"
keywords = ["cli", "unicode", "art", "render", "terminal"]
license = { file = "LICENSE" }

[project.urls]
"Homepage" = "https://github.com/ultraflame4/CliR"
"Bug Tracker" = "https://github.com/ultraflame4/CliR/issues"

[project.scripts]
clirender = "CliRenderer.__main__:main"

#[build-system]
#build-backend = "setuptools.build_meta"
#requires = ["setuptools", "wheel"]

[tool.setuptools]
packages = ["CliRenderer"]

[tool.poetry]
name = "CliRenderer"
version = "0.1.6"
description = ""
authors = ["ultr42 <ultraflame4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
numpy = "~1.26.1"
Pillow = "^10.1.0"
rich = "~13.6.0"
typer = "~0.9.0"
imageio = "~2.31.5"
validators = "~0.22.0"
setuptools = "^68.2.2"


[build-system]
requires = ["poetry-core","setuptools"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.build]
script = "build.py"
generate-setup-file = true